Fix reference to old github repository
This repository was renamed, but the old bcmi-labs version was still
used in the code. Since github redirects renamed repositories, things
still worked if you cloned from the new repo, but you would end up with
two copies of the repository, with the "auth" subpackage being used from
the other copy.