Change TestDockerPsApi checking real docker ps -a and mqtt msg
In this case I use to check real docker ps -a command called with exec
and msg from mqtt, I want check len of array of docker container in
machine. Using this we are uncoupled from status of machine container
but the focus is on correctly respond by production code of application.

Signed-off-by: Federico Guerinoni <fguerinoni@develer.com>